Subo: - infraestructura preparada para recolectar los datos de las máquinas virtuales activas y la distribución de las máquinas virtuales. - formatos de paquete modificados para utilizar ints en lugar de longs (y así ser coherentes con lo que hay en la base de datos). - pruebas unitarias de los módulos de la base de datos revisadas para incluir los nuevos métodos. - testers modificados para manipular los nuevos paquetes. - servidor de máquinas virtuales dummy modificado para incorporar la nueva funcionalidad.
 On branch infrastructure-upstream
 Changes to be committed:

	modified:   src/CygnusCloud/clusterServer/connector/clusterServerConnector.py
	modified:   src/CygnusCloud/clusterServer/networking/packets.py
	modified:   src/CygnusCloud/clusterServer/reactor/clusterServerReactor.py
	modified:   src/CygnusCloud/database/systemStatusDB/systemStatusDBReader.py
	modified:   src/CygnusCloud/database/systemStatusDB/systemStatusDBWriter.py
	modified:   src/CygnusCloud/database/tests/systemStatusDBTests.py
	modified:   src/CygnusCloud/database/tests/vmServerDBTests.py
	modified:   src/CygnusCloud/database/vmServer/runtimeData.py
	modified:   src/CygnusCloud/testing/clusterServerTester.py
	modified:   src/CygnusCloud/testing/dummyVMServer.py
	modified:   src/CygnusCloud/virtualMachineServer/packets.py
	modified:   src/CygnusCloud/virtualMachineServer/vmClient.py

+++ src/CygnusCloud/virtualMachineServer/packets.py
@@ -28,8 +28,8 @@ class VMServerPacketHandler(object):
         """
         p = self.__packetCreator.createPacket(5)
         p.writeInt(VM_SERVER_PACKET_T.CREATE_DOMAIN)
-        p.writeLong(machineId)
-        p.writeLong(userId)
+        p.writeInt(machineId)
+        p.writeInt(userId)
         return p
     
     def createVMConnectionParametersPacket(self, userID, vncServerIP, vncServerPort, password):
@@ -44,7 +44,7 @@ class VMServerPacketHandler(object):
         """
         p = self.__packetCreator.createPacket(5)
         p.writeInt(VM_SERVER_PACKET_T.DOMAIN_CONNECTION_DATA)
-        p.writeLong(userID)
+        p.writeInt(userID)
         p.writeString(vncServerIP)
         p.writeInt(vncServerPort)
         p.writeString(password)
@@ -110,7 +110,7 @@ class VMServerPacketHandler(object):
         p.writeInt(sequenceSize)
         p.writeString(serverIPAddress)
         for row in data :
-            p.writeLong(row["UserID"])
+            p.writeInt(row["UserID"])
             p.writeInt(row["VMID"])
             p.writeString(row["VMName"])
             p.writeInt(row["VNCPort"])
@@ -130,10 +130,10 @@ class VMServerPacketHandler(object):
         packet_type = p.readInt()
         result["packet_type"] = packet_type
         if (packet_type == VM_SERVER_PACKET_T.CREATE_DOMAIN) :
-            result["MachineID"] = p.readLong()
-            result["UserID"] = p.readLong()
+            result["MachineID"] = p.readInt()
+            result["UserID"] = p.readInt()
         elif (packet_type == VM_SERVER_PACKET_T.DOMAIN_CONNECTION_DATA):
-            result["UserID"] = p.readLong()
+            result["UserID"] = p.readInt()
             result["VNCServerIP"] = p.readString()
             result["VNCServerPort"] = p.readInt()
             result["VNCServerPassword"] = p.readString()
